Basic Win32 support:
* Stubs for GetPort*/SetPort*

* OpenPort has been implemented. It opens the serial port in
overlapped mode, retrieves the initial settings, changes to binary. It
also calls SetCommMask to listen for BREAK, RXCHAR, and TXEMTPY
events. Finally, a corresponding event is created.

* ClosePort cancels the overlapped operation, restores the initial
port settings, and closes the device.

* SercdSelect has been implemented as well. Tricky. It starts out with
firing an overlapped WaitCommEvent. If an immediate result is not
available, it's event is added to list of events to wait for. We are
also waiting for socket events.

On device events, the global variables DeviceReadChars and
DeviceWritable are updated.

For sockets, WSAEnumNetworkEvents is used to check events both on the
listenting and connection socket.

* The WriteToDev/ReadFromDev/WriteToNet/ReadFromNet functions have
been implemented as well.